changed the MaxShootewrSpeed from an int to a double

+++ src/main/java/frc/robot/Constants.java
@@ -86,7 +86,7 @@ public final class Constants {
     //PID
     public static double BallPIDTolerance = 0d;
     
-    public static int MaxShooterSpeed = 1;
+    public static double MaxShooterSpeed = 1;
 	
     /*----------ClimbSystem----------*/
